requirements_dev.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements_dev.txt --all-extras
annotated-doc==0.0.3
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   openai
    #   starlette
azure-ai-inference==1.0.0b9
    # via git-sms (pyproject.toml)
azure-core==1.36.0
    # via azure-ai-inference
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   tqdm
coverage==7.11.2
    # via pytest-cov
distro==1.9.0
    # via openai
fastapi==0.121.1
    # via git-sms (pyproject.toml)
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.3.0
    # via pytest
isodate==0.7.2
    # via azure-ai-inference
jiter==0.11.1
    # via openai
openai==2.7.1
    # via git-sms (pyproject.toml)
packaging==25.0
    # via pytest
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
pydantic==2.12.4
    # via
    #   fastapi
    #   openai
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via pytest
pytest==9.0.0
    # via
    #   git-sms (pyproject.toml)
    #   pytest-cov
    #   pytest-mock
pytest-cov==7.0.0
    # via git-sms (pyproject.toml)
pytest-mock==3.15.1
    # via git-sms (pyproject.toml)
python-dotenv==1.2.1
    # via git-sms (pyproject.toml)
python-multipart==0.0.20
    # via git-sms (pyproject.toml)
regex==2025.11.3
    # via tiktoken
requests==2.32.5
    # via
    #   git-sms (pyproject.toml)
    #   azure-core
    #   tiktoken
ruff==0.14.4
    # via git-sms (pyproject.toml)
sniffio==1.3.1
    # via
    #   anyio
    #   openai
starlette==0.49.3
    # via fastapi
tiktoken==0.12.0
    # via git-sms (pyproject.toml)
tqdm==4.67.1
    # via openai
typing-extensions==4.15.0
    # via
    #   anyio
    #   azure-ai-inference
    #   azure-core
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.5.0
    # via requests
uvicorn==0.38.0
    # via git-sms (pyproject.toml)
